mv clust rotation tests
